+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.7.0] - 2022-11-18
 ### Added
+- fwcli command `standard:services:add` for add new standard services.
 - Keepalived plugin.
 - Suricata plugin.
 - Zeek plugin.